backend/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
python-dateutil==2.8.2
schedule==1.2.0
